+++ build_request_form.go
@@ -12,6 +12,12 @@ import (
 	"github.com/Alsond5/gogram/models"
 )
 
+type inputMediaOrPaidMedia interface {
+	GetData() io.Reader
+	GetMedia() string
+	MarshalInputMedia() ([]byte, error)
+}
+
 func writeParamsToForm[T any](form *multipart.Writer, params *T) error {
 	v := reflect.ValueOf(params).Elem()
 
@@ -33,12 +39,28 @@ func writeParamsToForm[T any](form *multipart.Writer, params *T) error {
 		var err error
 
 		switch vv := v.Field(i).Interface().(type) {
+		case models.InputMedia:
+			err = addFormFieldInputMedia(form, fieldName, vv)
 		case string:
 			err = addFormFieldString(form, fieldName, vv)
 		case *models.InputFileUpload:
 			err = addFormFieldInputFileUpload(form, fieldName, vv)
 		case *models.InputFileString:
 			err = addFormFieldString(form, fieldName, vv.Data)
+		case []models.InputMedia:
+			var ss []inputMediaOrPaidMedia
+			for _, m := range vv {
+				ss = append(ss, m)
+			}
+
+			err = addFormFieldInputMediaSlice(form, fieldName, ss)
+		case []models.InputPaidMedia:
+			var ss []inputMediaOrPaidMedia
+			for _, m := range vv {
+				ss = append(ss, m)
+			}
+
+			err = addFormFieldInputMediaSlice(form, fieldName, ss)
 		default:
 			err = addFormFieldDefault(form, fieldName, vv)
 		}
@@ -49,6 +71,61 @@ func writeParamsToForm[T any](form *multipart.Writer, params *T) error {
 	}
 
 	return nil
+}
+
+func addFormFieldInputMediaSlice(form *multipart.Writer, fieldName string, value []inputMediaOrPaidMedia) error {
+	var fields []string
+	for _, vv := range value {
+		data, err := addFormFieldInputMediaItem(form, vv)
+		if err != nil {
+			return err
+		}
+
+		fields = append(fields, string(data))
+	}
+
+	w, err := form.CreateFormField(fieldName)
+	if err != nil {
+		return err
+	}
+
+	_, err = io.Copy(w, strings.NewReader("["+strings.Join(fields, ",")+"]"))
+	return err
+}
+
+func addFormFieldInputMediaItem(form *multipart.Writer, vv inputMediaOrPaidMedia) ([]byte, error) {
+	if !strings.HasPrefix(vv.GetMedia(), "attach://") {
+		return vv.MarshalInputMedia()
+	}
+
+	filename := strings.TrimPrefix(vv.GetMedia(), "attach://")
+	field, err := form.CreateFormFile(filename, filename)
+
+	if err != nil {
+		return nil, err
+	}
+
+	_, err = io.Copy(field, vv.GetData())
+	if err != nil {
+		return nil, err
+	}
+
+	return vv.MarshalInputMedia()
+}
+
+func addFormFieldInputMedia(form *multipart.Writer, fieldName string, vv inputMediaOrPaidMedia) error {
+	data, err := addFormFieldInputMediaItem(form, vv)
+	if err != nil {
+		return err
+	}
+
+	w, err := form.CreateFormField(fieldName)
+	if err != nil {
+		return err
+	}
+
+	_, err = io.Copy(w, bytes.NewReader(data))
+	return err
 }
 
 func addFormFieldInputFileUpload(form *multipart.Writer, fieldName string, value *models.InputFileUpload) error {