+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.3.0 (12-09-2025)
+
+- Added support for `InputMedia` (photo, video, animation, document & audio) in Telegram bot API
+- Added support for `InputPaidMedia` (photo & video) in Telegram bot API
+- Implemented handling of slices for both `InputMedia` and `InputPaidMedia`
+
 ## v0.2.0 (11-09-2025)
 
 - Support media upload in form building